% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.tree-conversion.R
\name{depth}
\alias{depth}
\title{Check depth of a list}
\usage{
depth(x)
}
\arguments{
\item{x}{list}
}
\value{
integer
}
\description{
Check depth of a list
}
\author{
Jasper Schelfhout \email{jasper.schelfhout@openanalytics.eu}
}
