% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.tree-conversion.R
\name{dfrapply}
\alias{dfrapply}
\title{Recursively apply function to all data.frames in a nested list}
\usage{
dfrapply(list, f, ...)
}
\arguments{
\item{list}{(nested) list containing data.frames}

\item{f}{function to apply to each data.frame}

\item{...}{extra arguments to f}
}
\value{
list
}
\description{
Recursively apply function to all data.frames in a nested list
}
\author{
Jasper Schelfhout \email{jasper.schelfhout@openanalytics.eu}
}
