% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.tree-conversion.R
\name{dfToTree}
\alias{dfToTree}
\title{Converts a data.frame  to a data.tree format}
\usage{
dfToTree(df, hierarchy = colnames(df))
}
\arguments{
\item{df}{data.frame}

\item{hierarchy}{ordered character vector of column names defining the hierarchy}
}
\value{
nested list
}
\description{
Converts a data.frame  to a data.tree format
}
\examples{
\dontrun{
df <- data.frame(Titanic)
tree <- dfToTree(df, c("Sex", "Class", "Survived"))
}
}
\author{
Jasper Schelfhout \email{jasper.schelfhout@openanalytics.eu}
}
