% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-selected.R
\name{get_selected_nodes}
\alias{get_selected_nodes}
\title{Get Selected Nodes}
\usage{
get_selected_nodes(tree = NULL, field = NULL)
}
\arguments{
\item{tree}{\code{List} returned from a 'ShinyTree'.}

\item{field}{\code{character} name of field to return instead of 'text'}
}
\description{
Get Selected Nodes
}
