% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-selected.R
\name{get_state_nodes}
\alias{get_state_nodes}
\title{Get State for Leaf Nodes}
\usage{
get_state_nodes(tree = NULL)
}
\arguments{
\item{tree}{\code{List} returned from a 'ShinyTree'.}
}
\description{
Get State for Leaf Nodes
}
