% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.tree-conversion.R
\name{nodesToDf}
\alias{nodesToDf}
\title{Convert the nodes of a tree into a data.frame}
\usage{
nodesToDf(tree, hierarchy)
}
\arguments{
\item{tree}{named nested list}

\item{hierarchy}{sorted character vector with name for each level of the list}
}
\value{
nested list with one level less that is stacked to a data.frame
}
\description{
Convert the nodes of a tree into a data.frame
}
\author{
Jasper Schelfhout
}
