% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-tree.R
\name{renderEmptyTree}
\alias{renderEmptyTree}
\title{Render an empty ShinyTree}
\usage{
renderEmptyTree()
}
\arguments{
\item{env}{The environment in which \code{expr} should be evaluated.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.
\code{\link{shinyTree}}.}
}
\description{
Renders a tree with no defined nodes.
}
\seealso{
\code{\link{shinyTree}}
}
