% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_traversal.R
\name{set_attr}
\alias{set_attr}
\title{Set attribute on object}
\usage{
set_attr(obj, attr_name, attr_value)
}
\arguments{
\item{obj}{any object}

\item{attr_name}{name of attribute to set}

\item{attr_value}{value of attribute to set}
}
\value{
the same object with set attribute
}
\description{
Set attribute on object
}
