% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-tree.R
\name{shinyTree}
\alias{shinyTree}
\title{Create a Shiny Tree}
\usage{
shinyTree(outputId, checkbox = FALSE, search = FALSE, dragAndDrop = FALSE,
  theme = "default", three_state = TRUE, whole_node = TRUE,
  tie_selection = TRUE)
}
\arguments{
  \item{outputId}{The ID associated with this element}
  \item{checkbox}{If \code{TRUE}, will enable checkboxes
  next to each node to make the selection of multiple nodes
  in the tree easier.} \item{search}{If \code{TRUE}, will
  enable search functionality in the tree by adding a search
  box above the produced tree. Alternatively, you can set the
  parameter to the ID of the text input you wish to use as the
  search field.} \item{dragAndDrop}{If \code{TRUE}, will allow
  the user to rearrange the nodes in the tree.} \item{theme}
  {jsTree theme, one of \code{default}, \code{default-dark},
  or \code{proton}. #' @param three_state If \code{TRUE},
  enables three-state checkboxes.} \item{whole_node}{If
  \code{TRUE}, clicking anywhere on the node acts as clicking
  on the checkbox.} \item{tie_selection}{If \code{TRUE},
  checkboxes are bound to the general tree selection.}
}
\description{
This creates a spot in your Shiny UI for a shinyTree which
can then be filled in using \code{\link{renderTree}}
}
\seealso{
\code{\link{renderTree}}
}

