% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.tree-conversion.R
\name{stackList}
\alias{stackList}
\title{Combine named list into single data.frame with extra column}
\usage{
stackList(list, name)
}
\arguments{
\item{list}{named list}

\item{name}{character}
}
\value{
data.frame
}
\description{
Combine named list into single data.frame with extra column
}
\author{
Jasper Schelfhout
}
