% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-tree.R
\name{tree}
\alias{tree}
\title{Create a Tree Widget Using the JavaScript Library jsTree}
\usage{
tree(data, options = list(), plugins = NULL, width = NULL,
  height = NULL)
}
\arguments{
\item{data}{A list with a tree-like structure, either in the standard or the
alternative format; see \url{https://www.jstree.com/docs/json/}.}

\item{options}{A list of initialization options, e.g., to control the looks
of jsTree; see \url{https://www.jstree.com/api/}.}

\item{plugins}{A list of plugins to enable further functionality of jsTree;
see \url{https://www.jstree.com/plugins/}.}

\item{width}{The width of the input container, e.g., \code{'100\%'}; see
\link{validateCssUnit}.}

\item{height}{The height of the input container, e.g., \code{'200px'}; see
\link{validateCssUnit}.}
}
\value{
A tree widget that can be added to a UI definition.
}
\description{
Create a tree widget using the JavaScript library jsTree.
}
\references{
See \url{https://www.jstree.com/} for the full documentation.
}
