% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs-shiny.R
\name{treeOutput}
\alias{treeOutput}
\alias{renderTree}
\title{Helper Functions for Using Shinytree in Shiny}
\usage{
treeOutput(outputId, width = "100\%", height = "100\%")

renderTree(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{The \code{output} slot that will be used to access the
values.}

\item{width}{The width of the input container, e.g., \code{'100\%'}; see
\link{validateCssUnit}.}

\item{height}{The height of the input container, e.g., \code{'200px'}; see
\link{validateCssUnit}.}

\item{expr}{An expression to create a tree widget (via \code{tree()}),
or a data object to be passed to \code{tree()} to create a tree widget.}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})?
This is useful if you want to save an expression in a variable.}
}
\description{
These functions are like most \code{fooOutput()} and \code{renderFoo()}
functions in the \pkg{shiny} package. The former is used to create a
container for a tree, and the latter is used in the server logic to
render the tree.
}
\examples{
\donttest{
inst/examples/tree.R
}

}
