% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.tree-conversion.R
\name{treeToDf}
\alias{treeToDf}
\title{Convert tree into data.frame}
\usage{
treeToDf(tree, hierarchy)
}
\arguments{
\item{tree}{named nested list}

\item{hierarchy}{sorted character vector with name for each level of the list}
}
\value{
data.frame
}
\description{
Convert tree into data.frame
}
\author{
Jasper Schelfhout
}
