% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.tree-conversion.R
\name{treeToJSON}
\alias{treeToJSON}
\title{Converts a data.tree to a JSON format}
\usage{
treeToJSON(tree, topLevelSlots = c("id", "text", "icon", "state",
  "li_attr", "a_attr"), pretty = FALSE)
}
\arguments{
\item{tree, }{the data.tree which should be parses}

\item{topLevelSlots, }{a vector of slot names which should not be stored in
the resulting data slot but on the top level of the node}

\item{pretty, }{logical. If \code{TRUE} the resulting JSON is prettified}
}
\value{
a JSON string representing the data.tree
}
\description{
Walk through a \code{\link{data.tree}} and constructs a JSON string,
which can be rendered by shinyTree. The JSON string generated follows the 
\href{https://www.jstree.com/docs/json}{jsTree specifications}. In particular 
it encodes children nodes via the \sQuote{children} slot.
}
\details{
All atomic or list slots of a node in the tree are stored in a data slot in 
the resulting JSON.

If the user wants to store some slots not in the data slot but on the top 
level of the node, parameter \code{topLevelSlots} can be used. This is useful
for additional parameters such as \sQuote{icon}, \sQuote{li_attr} or 
\sQuote{a_attr}, which jsTree expect to be on the top level of the node.

An example of how to make use of this functionality can be found in the
example folder of this library.
}
\section{Note}{

\code{\link{updateTree}} and \code{\link{renderTree}} need an unevaluated JSON 
string. Hence, this function returns a string rather than the JSON object itself.
}

\author{
Thorn Thaler, \email{thorn.thaler@thothal.at}
}
