% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-tree.R
\name{updateTree}
\alias{updateTree}
\title{Update the tree with new data}
\usage{
updateTree(session, treeId, data = NULL)
}
\arguments{
\item{session}{The current session variable.}

\item{treeId}{The identifier for the shinyTree object}

\item{data}{JSON data or nested list representing the new tree structure.}
}
\description{
Extract the nodes from the tree that are selected in a more
convenient format. You can choose which format you prefer.
}
