% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-tree.R
\name{renderEmptyTree}
\alias{renderEmptyTree}
\title{Render an empty ShinyTree}
\usage{
renderEmptyTree()
}
\description{
Renders a tree with no defined nodes.
\code{\link{shinyTree}}.
}
\seealso{
\code{\link{shinyTree}}
}
