% Generated by roxygen2 (4.0.1): do not edit by hand
\name{shinyTree}
\alias{shinyTree}
\title{Create a Shiny Tree}
\usage{
shinyTree(outputId, checkbox = FALSE, selected = NULL)
}
\arguments{
  \item{outputId}{The ID associated with this element}

  \item{checkbox}{If \code{TRUE}, will enable checkboxes
  next to each node to make the selection of multiple nodes
  in the tree easier.}

  \item{selected}{The input ID associated with tree's
  currently selected elements. The names of the node(s)
  that are currently selected in the tree will be available
  in an input by this name.}
}
\description{
This creates a spot in your Shiny UI for a shinyTree which
can then be filled in using \code{\link{renderTree}}
}
\seealso{
\code{\link{renderTree}}
}

