% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-tree.R
\name{shinyTree}
\alias{shinyTree}
\title{Create a Shiny Tree}
\usage{
shinyTree(outputId, checkbox = FALSE, search = FALSE,
  dragAndDrop = FALSE, types = NULL, theme = "default",
  themeIcons = TRUE, themeDots = TRUE)
}
\arguments{
\item{outputId}{The ID associated with this element}

\item{checkbox}{If \code{TRUE}, will enable checkboxes next to each node to 
make the selection of multiple nodes in the tree easier.}

\item{search}{If \code{TRUE}, will enable search functionality in the tree by adding
a search box above the produced tree. Alternatively, you can set the parameter
to the ID of the text input you wish to use as the search field.}

\item{dragAndDrop}{If \code{TRUE}, will allow the user to rearrange the nodes in the
tree.}

\item{types}{enables jstree types functionality when sent proper json (please see the types example)}

\item{theme}{jsTree theme, one of \code{default}, \code{default-dark}, or \code{proton}.}

\item{themeIcons}{If \code{TRUE}, will show theme icons for each item.}

\item{themeDots}{If \code{TRUE}, will include level dots.}
}
\description{
This creates a spot in your Shiny UI for a shinyTree which can then be filled
in using \code{\link{renderTree}}
}
\seealso{
\code{\link{renderTree}}
}
